requirements.txt
flask
flask-sqlalchemy
psycopg2-binary
flask-migrate
flask-socketio
flask-session
flask-script
python-socketio
eventlet
pandas
openpyxl
werkzeug
Flask-Mail
gunicorn
python-dotenv
xlsxwriter
flask-login
numpy
flask-jwt-extended